Put enums into new file
