+++ ManageCoursesUi.Tests/Helpers/TestHelper.cs
@@ -2,19 +2,10 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using GovUk.Education.ManageCourses.ApiClient;
+using ManageCoursesUi.Tests.Enums;
 
 namespace ManageCoursesUi.Tests.Helpers
 {
-    /// <summary>
-    /// Defines the type of data that needs to be generated by the helper class
-    /// </summary>
-    public enum EnumDataType
-    {
-        SingleVariantOneMatch,//data setup with one matching variant
-        MultiVariantOneMatch,//data setup with multiple variants and only one metching variant
-        SingleVariantNoMatch,//data setup with one none matching variant
-        MultiVariantNoMatch//data setup with multiple variants none matching
-    }
     /// <summary>
     /// This helper class generates the exact structure of the data returned from the real Api.
     /// It also generates non-targeted data so that the end point will error
